chore: refactored files/ folder (#13)
