+++ .gitignore
@@ -57,7 +57,7 @@ _templates
 /public/favicons/*
 
 # catalog files
-/files/**/out/**
+catalog/**
 
 # python
 __pycache__/
