+++ catalog-build/build-ncpi-catalog.ts
@@ -1,6 +1,6 @@
-import { PlatformStudy } from "../../app/apis/catalog/ncpi-catalog/common/entities";
-import { parseContentRows, readFile } from "../../app/utils/tsvParser";
-import { writeAsJSON } from "../common/utils";
+import { PlatformStudy } from "../app/apis/catalog/ncpi-catalog/common/entities";
+import { parseContentRows, readFile } from "../app/utils/tsvParser";
+import { writeAsJSON } from "./common/utils";
 import { buildNCPICatalogPlatforms } from "./build-plaftorms";
 import { buildNCPIPlatformStudies } from "./build-platform-studies";
 import { SOURCE_FIELD_KEY, SOURCE_FIELD_TYPE, tsvPath } from "./constants";
@@ -31,12 +31,12 @@ async function buildCatalog(): Promise<void> {
   const ncpiCatalogPlatforms = buildNCPICatalogPlatforms(ncpiPlatformStudies);
 
   await writeAsJSON(
-    "ncpi-catalog/out/ncpi-platform-studies.json",
+    "../catalog/ncpi-platform-studies.json",
     Object.fromEntries(ncpiPlatformStudies.entries())
   );
 
   await writeAsJSON(
-    "ncpi-catalog/out/ncpi-platforms.json",
+    "../catalog/ncpi-platforms.json",
     Object.fromEntries(ncpiCatalogPlatforms.entries())
   );
 }
