+++ catalog-build/common/disease_abbrev_mapping.tsv
