+++ catalog-build/constants.ts
@@ -1,4 +1,4 @@
-export const tsvPath = "ncpi-catalog/files/dashboard-source-ncpi.tsv";
+export const tsvPath = "source/dashboard-source-ncpi.tsv";
 
 export enum Platform {
   ANVIL = "AnVIL",
