+++ catalog-build/package.json
@@ -5,11 +5,11 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build-ncpi-db": "esrun ncpi-catalog/build-ncpi-catalog.ts",
-    "update-anvil-source": "esrun ncpi-catalog/update-anvil-source.ts",
-    "update-bdc-source": "esrun ncpi-catalog/update-bdc-source.ts",
-    "update-crdc-source": "esrun ncpi-catalog/update-crdc-source.ts",
-    "update-kfdrc-source": "esrun ncpi-catalog/update-kfdrc-source.ts",
+    "build-ncpi-db": "esrun build-ncpi-catalog.ts",
+    "update-anvil-source": "esrun update-anvil-source.ts",
+    "update-bdc-source": "esrun update-bdc-source.ts",
+    "update-crdc-source": "esrun update-crdc-source.ts",
+    "update-kfdrc-source": "esrun update-kfdrc-source.ts",
     "update-all-ncpi-sources": "npm run update-crdc-source && npm run update-bdc-source && npm run update-anvil-source && npm run update-kfdrc-source"
   },
   "author": "",
