+++ catalog-build/update-crdc-source.ts
@@ -1,5 +1,5 @@
 import fetch from "node-fetch";
-import { DbGapId } from "../../app/apis/catalog/ncpi-catalog/common/entities";
+import { DbGapId } from "../app/apis/catalog/ncpi-catalog/common/entities";
 import { Platform } from "./constants";
 import {
   getPlatformStudiesStudyIds,
